Port to Python3/GTK+3
This doesn't work all that well (not yet?). There is a bytes/str issue
with configparser. More importantly, it now crashes with a segfault
whenever an eject is performed, most likely a GTK3 issue. Add to that,
the fact that now GtkStatusIcon (aka tray icons) are deprecated in GTK3,
and the tray icon is an important part of gmountman, I think that this
will stay with GTK2 for the foreseeable future.